refactor: split base_alerts.py into different modules
AnomalyAlertController was put in anomaly_alerts.py file

StaticEventAlertController was put in event_alerts.py file

StaticKpiAlertController was put in static_kpi_alerts.py file

Free functions in base_alerts.py file were put in utils.py file

Entry point functions were put in __init__.py file
