+++ chaos_genius/alerts/static_kpi_alerts.py
@@ -1,0 +1,6 @@
+class StaticKpiAlertController:
+    def __init__(self, alert_info):
+        self.alert_info = alert_info
+
+    def check_and_prepare_alert(self):
+        pass
