+++ chaos_genius/alerts/utils.py
@@ -2,7 +2,7 @@
 
 import datetime
 import heapq
-from typing import Dict, List, Tuple
+from typing import Dict, List, Tuple, Union
 
 from chaos_genius.alerts.constants import (
     ALERT_DATETIME_FORMAT,
@@ -53,3 +53,49 @@ def count_anomalies(points: List[Dict]) -> Tuple[int, int]:
     overall = sum(1 for point in points if point["Dimension"] == "Overall KPI")
     subdims = total - overall
     return overall, subdims
+
+
+def change_message_from_percent(percent_change: Union[str, int, float]) -> str:
+    """Creates a change message from given percentage change.
+
+    percent_change will be:
+        - "-" in case the last data point was missing
+        - 0 (int) in case there was no change
+        - positive value (int/float) in case there was an increase
+        - negative value (int/float) in case there was a decrease
+    """
+    if isinstance(percent_change, str):
+        return "-"
+    elif percent_change == 0:
+        return "No change (-)"
+    elif percent_change > 0:
+        return f"Increased by ({percent_change}%)"
+    else:
+        return f"Decreased by ({percent_change}%)"
+
+
+def _format_anomaly_points(points: List[dict]):
+    for anomaly_point in points:
+        anomaly_point["series_type"] = (
+            "Overall KPI"
+            if anomaly_point.get("anomaly_type") == "overall"
+            else anomaly_point["series_type"]
+        )
+        for key, value in anomaly_point.items():
+            if key in ANOMALY_TABLE_COLUMNS_HOLDING_FLOATS:
+                anomaly_point[key] = round(value, 2)
+        if anomaly_point["series_type"] != "Overall KPI":
+            anomaly_point["series_type"] = convert_query_string_to_user_string(
+                anomaly_point["series_type"]
+            )
+
+
+def find_percentage_change(
+    curr_val: Union[int, float], prev_val: Union[int, float]
+) -> Union[int, float, str]:
+    """Calculates percentage change between previous and current value."""
+    if prev_val == 0:
+        return "-"
+    change = curr_val - prev_val
+    percentage_change = (change / prev_val) * 100
+    return round_number(percentage_change)
